refactor: Rename `…ins/aws/package/lib/generateArtifactDirectoryName.js`
