+++ lib/plugins/aws/package/index.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const mergeCustomProviderResources = require('./lib/mergeCustomProviderResources');
 const stripNullPropsFromTemplateResources = require('./lib/stripNullPropsFromTemplateResources');
-const generateArtifactDirectoryName = require('./lib/generateArtifactDirectoryName');
+const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
 const generateCoreTemplate = require('./lib/generateCoreTemplate');
 const saveServiceState = require('./lib/saveServiceState');
 const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
