+++ lib/plugins/aws/package/lib/generate-artifact-directory-name.js
