+++ test/unit/lib/plugins/aws/package/lib/generate-artifact-directory-name.test.js
